Fix fences being put on roads
- The issue was caused by the assumption that tiles can be
changed multiple times (at the same coords) in a single update call
It may be important to note here that this is not the case, because
the Tile GO after the first setTile call is in the add queue, and
remove doesn't seem to change this. At the end of the update call
both tiles get added to the scene.
This shouldn't cause any issues during normal play, as this is the
equivalent of adding and removing the same GO during an update.