Add QuadTree and some helper classes
Added a generic QuadTree implementation to the engine for spatial
hashing and, in turn, faster collision detection.

It relies on some other helper classes which are now part of the
engine's Helper namespace and are completely quadtree-independent so
that they can be reused anywhere:

ObjectPool<T> can be used for pooling heap-based objects that are
created frequently and temporarily, to reduce memory allocation time.

Vectangles are float-based rectangles, following an almost 1:1
implementation to MonoGame's Rectangle (functionally at least).